api/pyproject.toml

[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["raulpy271 <raulrufino0@gmail.com>"]

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.group.dev]
optional = true

[tool.poetry.dependencies]
python = "^3.12"
redis = "^5.2.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.43"}
psycopg = {extras = ["binary"], version = "^3.2.3"}
pyjwt = "^2.10.1"
fastapi = {extras = ["standard"], version = "^0.116.1"}
azure-storage-blob = "^12.24.1"
azure-identity = "^1.19.0"
python-multipart = "^0.0.20"
aiohttp = "^3.11.12"
python-dateutil = "^2.9.0.post0"
sqlmodel = "^0.0.24"
pydantic = "^2.11.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
python-dotenv = "^1.0.1"
fakeredis = "^2.26.2"
httpx = "^0.28.1"
mimesis = "^18.0.0"
black = "^25.1.0"
flake8 = "^7.1.2"
alembic = "^1.16.2"
pre-commit = "^4.2.0"
aiosqlite = "^0.21.0"

[tool.black]
target-version = ["py312"]

[tool.pytest.ini_options]
testpaths = "tests"
usefixtures = "general"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
addopts = [
    "--import-mode=importlib",
]

